refactor: Billing::OpaqueDataPaymentToken -> Gateways::AuthorizeNet::PaymentToken
