+++ lib/active_merchant/billing.rb
@@ -10,7 +10,6 @@ require 'active_merchant/billing/base'
 require 'active_merchant/billing/check'
 require 'active_merchant/billing/payment_token'
 require 'active_merchant/billing/apple_pay_payment_token'
-require 'active_merchant/billing/opaque_data_payment_token'
 require 'active_merchant/billing/response'
 require 'active_merchant/billing/gateways'
 require 'active_merchant/billing/gateway'
